fix: Keep new checks UI code inside the ESLint budget
CI failed on the ESLint budget check. The checks feature added
warnings that pushed three rules over their baselines:

- WorkOrderChecksSection.tsx exported LEVEL_LABEL and formatCheckScore
  next to components, which broke react-refresh/only-export-components.
  Move both to lib/workOrderChecks.ts.
- presentWorkOrderCheck had a cyclomatic complexity of 18. Extract the
  level and empty-string mapping into small helpers.
- WorkOrderActivityTimeline.tsx and useFactoryData.ts crossed the
  500-line cap. Move CheckReportedEventBody and ArtifactEventBody to
  the timeline/ folder next to the other timeline items, and move
  useWorkOrderChecks into its own hook file.
